Install Rust on CircleCI
We install Rust using rustup which is the most common way to do this.
This gets us the current stable version (updated if necessary). I would
like to use the Debian package for this, but Rust is not stable enough
for that yet. The language is evolving, the libraries adopt new compiler
features, and the Rust team supports only the latest stable version.
Therefore we go for downloading stuff from the Internet via rustup.rs.

In addition to core Rust toolchain we install a couple of tools that we
use for quality assurance: code formatter, static analyzer, and dead
link checker. "deadlinks" take a while to compile and install, sadly.

In general, Rust tooling *looooves* to pull lots of stuff from the net
so we cache the downloaded packages and installed binaries if possible.
This cuts down compilation time considerably, but there's still way to
go. We do a clean build of Themis so we take our time to compile all
the dependencies of dependencies...

Aside from rustc we install a couple of native dependencies. pkg-config
is required to locate Themis after installation. Clang is a dependency
of bindgen crate which is used to generate bindings for Themis.