Run rust-themis tests during CI builds
Add "test_rust" target to Makefile which runs rust-themis tests similar
to other language wrappers. We skip the tests if Rust does not seem to
be installed on the system.

The test suite may be a bit large. We'll see how much time it takes
to run on the CI. It replicates the current setup used by the original
rust-themis on TravisCI [1]

- Disallow any warning by any tool (the compiler and alike)
- Ensure consistent code formatting with rustfmt
- Avoid common mistakes by running static analysis with Clippy
- Check both dynamic and static linkage to core Themis library
- Compile and run automated test suite for all packages
- Check that API documentation is likely to build on docs.rs
- Check that documentation does not contain dead URLs

[1]: https://github.com/ilammy/rust-themis/blob/master/.travis.yml