Install pkg-config files (#343)
* Install pkg-config files

pkg-config is an amazing tool for locating installed libraries and
generating linkage information. It makes it trivial to link any
program to a library. In the simplest case of a C program building
with make one just adds these lines to their Makefile:

    CFLAGS += $(pkg-config --cflags libthemis)
    LDFLAGS += $(pkg-config --libs libthemis)

And it just works. No need for CMake to probe for libraries, etc.

pkg-config is also able to generate correct flags for static linkage.
Static linkage is not trivial because of transitive dependencies.
For example, when linking to Themis dynamically it is sufficient
to add just `-lthemis` to linker's command line. However, for static
linkage one needs to tell `-static -lthemis -lsoter -lcrypto`,
explicitly including Soter (dependency of Themis) and SomethingSSL
(dependency of Soter) into the command line. pkg-config tracks
library dependencies and outputs correct flags as necessary.

Output of pkg-config is pretty straighforward so it may be used
not only via directly embedding into makefiles for C programs,
but by other ecosystems as well. It provides a consistent way
to locate native libraries, portable across all major operating
systems (yes, pkg-config is available on Windows as well).

Therefore it would be nice for Themis to include *.pc files for
its libraries. Each library binary should have its own pkg-config
file. These files are usually installed into "pkgconfig" directory
under the "lib" directory with library binaries. System directories
like /usr/lib/pkgconfig and /usr/local/lib/pkgconfig are usually
in the default search path of pkg-config.

The *.pc files are installed during "make install" phase and are
included into DEB and RPM packages.

We generate *.pc files during the build to include the PREFIX value
and version information there. For Soter we also include the exact
library set that's used by the selected cryptographic engine. It's
different for OpenSSL/LibreSSL and BoringSSL.

Note that for OpenSSL we could have used `Requires` directive as
OpenSSL is usually available via pkg-config as "libcrypto". However,
Themis build system does not use pkg-config to locate OpenSSL so
we do not rely on pkg-config for resolving dependencies. This is
actually important for builds on macOS with Homebrew which should
ensure that Homebrew's OpenSSL is used which is not present in the
default search path of pkg-config.

Also note that we do not include CFLAGS for Soter. That's because
Soter public header files do *not* expose the dependency on OpenSSL
which is really good design and separation of concerns, promoting
stability of Soter's ABI.

We use sed to perform template substitution. It's good enough (until
someone tries using exclamation marks `!` in their paths, but let's
not think about those weirdos). Note that we use shell pipes, not
`-i` flag of sed. That's because `-i` is actually GNU sed specific,
BSD systems usually do not have it.

* Make sure the output directory exists

BIN_PATH is normally created when building object files (grep "$(@D)"
in Makefiles), but *.pc files do not have any dependencies and thus
make decides to build these targets early. Well, create the output
directories early then.