Return SOTER_INVALID_PARAMETER if the caller tries to reuse hash context after finalising it for BoringSSL backend
In BoringSSL EVP_DigestFinal "cleans up" the internal hash context and fills the
whole structure with zeroes. We can use this as a marker that the context has
been finalised. We just access the first "digest" member, which is a pointer
to a specific digest algorithm implementation. If the pointer is NULL, it means
there is no algorithm associated with the context, therefore it was either
finalised or never initialised in the first place.