Don't use empty keys in Secure Cell examples
Secure Cell does not allow empty keys. Encryption and decryption will
fail if an empty key is used, but this will be detected later, not at
the construction site.

Currently we don't do anything about it (so just update the API docs),
but later this will become an enforced assertion.