Context Imprint mode requires non-empty context
Secure Cell *requires* non-empty user context when in context imprint
mode. It fails to operate if the provided context is empty. Currently
this is noted only in the very last example in the module-level
documentation. Let's be more explicit about this.

Later we'll get even more serious and add some assertions in the code
as well as a "Panics" section in the docs.