Update Rust binding to v0.0.3 (#349)
* Update Rust binding to v0.0.3

This is a cumulative update of rust-themis to version 0.0.3 from here:

    https://github.com/ilammy/rust-themis

You can see `git diff v0.0.2..v0.0.3` there for detailed changes.
Obviously, the diff does not apply as is and has been manually merged
into this repository, taking into account previous changes and the new
code layout.

Some notable changes since v0.0.2:

- Secure zeroing of key material is included into this update.

- Rust 2018 edition support and API documentation updates
  introduce a lot of diff clutter. Deal with it.

- Vendored build is back. It is mostly intended as a hack for
  docs.rs support rather than an actual use-case.

  This brings a new `libthemis-src` crate to contain the whole
  source code of Themis. It is implemented with a symbolic link
  to the top of the repository. Hopefully, this loop does not
  break anything anywhere...

  For one, symlink usage *has* broken `libthemis-src` compilation
  (due to a bug in 3rd-party crate) so vendored build is currently
  failing. We'll fix this issue later.

- There are some other interesting TODOs added and removed during
  the development. Just grep for them in you're interested.

These changes do not integrate rust-themis into the main build system.
It is possible to build rust-themis alone with Cargo, but the Makefile
does not know about us yet. This will be added later.

* Don't use empty keys in Secure Cell examples

Secure Cell does not allow empty keys. Encryption and decryption will
fail if an empty key is used, but this will be detected later, not at
the construction site.

Currently we don't do anything about it (so just update the API docs),
but later this will become an enforced assertion.

* Context Imprint mode requires non-empty context

Secure Cell *requires* non-empty user context when in context imprint
mode. It fails to operate if the provided context is empty. Currently
this is noted only in the very last example in the module-level
documentation. Let's be more explicit about this.

Later we'll get even more serious and add some assertions in the code
as well as a "Panics" section in the docs.