Fix vendoring of Themis source code
libthemis-src needs to embed the source code of core Themis library.
Previously, we did this by using git submodules when rust-themis was
developed in a separate repository. It was... satisfactory. (Not great
because "cargo package" refused to package "themis" directory as it
contained a Cargo.toml file.)

git submodule has been converted into a symlink to the root directory
when rust-themis was merged into the main repository. This should have
worked well, but it didn't. First of all, the "copy_dir" crate does not
copy symlinks and this breaks the build. Second, the root directory now
contains Cargo's target directory which is huge and copying it as well
significantly increases build times.

In order to reduce build time let's copy only the essential files.
Replace the quick-and-dirty "themis" symlink with a more refined
selection of files necessary to compile core Themis library. We don't
really need all the wrappers, documentation, etc. Just copy the source
code, the build system, and some other files which are unconditionally
included by the top-level makefile. (This should also significantly
reduce the size of the crate.)

Next, drop the "copy_dir" crate and replace it with our ad-hoc utility
for copying files that handles symlinks the way we need it. It's not
a general-purpose utility but it does its job well enough.

And while we're here, redirect output of "make install" to /dev/null.
It's annoying to look at when we run tests. Ideally we would like to
redirect it into stderr, but that's not possible with Command's API.
(We keep the original stderr though, so we'll see compilation errors
just fine.)

Unfortunately, running "cargo package" for libthemis-src is still kinda
broken because Cargo does not handle symlinks. This is a known issue.
Currently we will have to manually copy the source directory to resolve
all the symlinks and just get a distributable files. Annoying, yes.