Drop secure zeroing test
It's actually not alright.

This test is flaky and is not really safe to run. It happens to behave
itself in single-threaded environment, but in reality the memory freed
from the keys is instantly reused by the allocator and does not contain
zeros anymore. It's not even guaranteed to be actually allocated.

I guess it's fine to trust zeroize to do its thing and do not verify
that memory is indeed zeroed out. Let's drop this test because flaky
CI builds are bad.