More robust tests for data corruption
Writing some random value at arbitrary offset sometimes is not enough
to break Secure Cell. Internally, Secure Cell uses randomized keys
so the ciphertext may actually happen to contain "42" at exactly that
offset. Or so it seems to be when the tests are run by CircleCI.

Try flipping the bits instead of writing a constant. We do so in
documentation tests and this should be a better way to simulate
corrupted data.