Clarify the source of "curl | sh" instructions
I'm not a fan of this kind of installation that downloads random stuff
from the Internet, but we'll have to live with it for a while, until
Rust is stable enough for disto-packaged versions not becoming outdated
after half a year.

Let's at least add a comment about the source of the spell below so
that it can be properly audited.