Avoid copying source code during libthemis-src build
It turns out that we can avoid copying Themis source code. The build
system supports out-of-source builds with BUILD_PATH environment
variable. Just point it into OUT_DIR and Cargo should be fine.

Remove all the copying utilities and blatant lies about Themis from
the comments.

We still keep the "optimized" symlinks instead of reverting back to
a single one pointing to the root directory of the repo. This way
it's more friendly towards crate packaging: "themis" directory won't
have a Cargo.toml inside it and the whole crate size is smaller.