Integration tools for Rust wrapper
Add tools written in Rust which are usable by integration tests.

These are simply command-line utilities with a particular interface.
They are somewhat similar to the examples from docs/rust/examples
(keygen in particular) and may also serve as examples of Rust-Themis
usage. The tools are even added to Cargo.toml as examples so that they
are at least compile-checked during every build.

Writing these tools turned out to be quite useful exercise for the API.
For example, they have revealed that Rust-Themis API for Secure Cell is
a bit different from other language wrappers. Currently we do nothing
about it and just leave TODOs in the code. We'll improve the API later.