Add Rust-Themis to integration tests
Well, this is a bit tricky... Currently the integration tests work only
for scripting languages like Python, or for Go which has extremely fast
compilation times and right tooling. Basically, the current setup runs
tests like this: "$interpreter tools/$language/$tool.$extension $args".
Rust tools can't support this without additional help.

To work around this we introduce a whole bunch of wrapper Shell scripts
which act as testing tools and then forward their arguments to actual
tool binaries. This way we don't have to break assumptions like that
every testing tool has a filename extension. (Also, note the quote
usage in the scripts. That's how it must be because some arguments may
contain spaces and empty strings.)

Rust is a compiled language and its compilation times are quite long.
Simply using "cargo run --example" for each tool adds about a second
to the startup time (for Cargo to check that the files did not change),
assuming that all tools have been already compiled. In order to have
a more streamlined testing process we precompile the tools and install
them to Cargo's binary directory (~/.cargo/bin) which is in PATH.

We check for these installed binaries in wrapper shell scripts and use
them if they are not available. Otherwise we call Cargo to compile and
run the tools for us, in case you don't want to install them (and you
really don't want to do this outside of a temporary Docker container).