Do not use fixed user context with Secure Cell
It is really unwise to have user context the same for all messages
handled by Secure Cell. The idea behind context is for it to be
associated with message content, but distinct and separable from it.
Therefore the context has be different for each message, not the
same for all of them.

With this in mind, drop SecureCell::with_key_and_context() constructor
and introduce encrypt_with_context() and decrypt_with_context() methods
for Secure Cell operation modes. These methods accept the user context
as their first argument. Regular encrypt() and decrypt() methods are
equivalent to using an empty context.

Note that Secure Cell always requires non-empty context in context
imprint mode thus this flavor has only *_with_context() methods.

While we're here, seize this chance to replace ad-hoc generics in
Secure Cell methods with "impl Trait" feature available in Rust 2018.
This improves readability, especially for token-protect mode. However,
this is also a breaking change as impl Trait does not allow explicit
disambiguation.

Finally, with these changes we no longer need KeyBytes::empty()
constructor. Drop it.