Improve invocation of Rust integration tools
It's not really nice to install binaries into user's home directly
so let's try to do better. This is still a hack, but it's cleaner.

Originally we had this setup:

- "rustthemis_examples_install" target defined in top-level Makefile
  called Cargo to install integration testing tools to ~/.cargo/bin

- integration tests were running "bash tools/rust/*.sh" scripts
  when asked to run Rust tools

- the shell script were running integration tools from ~/.cargo/bin

This is bad because the developers can accidentally install tools
into their home directory.

Now we have the following hack:

- "rustthemis_integration_tools" target defined in test.mk file
  calls Cargo to install integration testing tools into Themis
  build directory ("build" at repository root by default)

- the tools are then copied to tools/rust with "*.rust" extension

- integration tests were run Rust tools as "env tools/rust/*.rust"
  pretending that they are actually scripts

This is still kinda bad because we pollute the tools/rust directory
with throwaway binaries. However, integration tests already do this
to tests/_integration directory with integration driver scripts, so
adding some more binaries should be acceptable.

(Oh, and the keygen tool has been officially renamed into keygen_tool
to avoid conflicts with docs/examples/rust/keygen.rs)