Consistent names of Secure Message methods
It does not really make sense to keep the wrap/unwrap naming scheme
since Secure Message API in Rust is different from the one in C.
Let's rename the methods to be consistent with documentation.

While we're here, update the methods to use impl Trait just like
we did before with Secure Cell.