Rust integration tests (#356)
* Integration tools for Rust wrapper

Add tools written in Rust which are usable by integration tests.

These are simply command-line utilities with a particular interface.
They are somewhat similar to the examples from docs/rust/examples
(keygen in particular) and may also serve as examples of Rust-Themis
usage. The tools are even added to Cargo.toml as examples so that they
are at least compile-checked during every build.

Writing these tools turned out to be quite useful exercise for the API.
For example, they have revealed that Rust-Themis API for Secure Cell is
a bit different from other language wrappers. Currently we do nothing
about it and just leave TODOs in the code. We'll improve the API later.

* Install Rust toolchain for integration tests

We're going to need Rust for integration tests so install it. We use
the same rustup.rs approach as for the main test suite. We also have
to install "clang" and "pkg-config" as they are dependencies of the
Rust-Themis build.

* Add Rust-Themis to integration tests

Well, this is a bit tricky... Currently the integration tests work only
for scripting languages like Python, or for Go which has extremely fast
compilation times and right tooling. Basically, the current setup runs
tests like this: "$interpreter tools/$language/$tool.$extension $args".
Rust tools can't support this without additional help.

To work around this we introduce a whole bunch of wrapper Shell scripts
which act as testing tools and then forward their arguments to actual
tool binaries. This way we don't have to break assumptions like that
every testing tool has a filename extension. (Also, note the quote
usage in the scripts. That's how it must be because some arguments may
contain spaces and empty strings.)

Rust is a compiled language and its compilation times are quite long.
Simply using "cargo run --example" for each tool adds about a second
to the startup time (for Cargo to check that the files did not change),
assuming that all tools have been already compiled. In order to have
a more streamlined testing process we precompile the tools and install
them to Cargo's binary directory (~/.cargo/bin) which is in PATH.

We check for these installed binaries in wrapper shell scripts and use
them if they are not available. Otherwise we call Cargo to compile and
run the tools for us, in case you don't want to install them (and you
really don't want to do this outside of a temporary Docker container).

* Improve invocation of Rust integration tools

It's not really nice to install binaries into user's home directly
so let's try to do better. This is still a hack, but it's cleaner.

Originally we had this setup:

- "rustthemis_examples_install" target defined in top-level Makefile
  called Cargo to install integration testing tools to ~/.cargo/bin

- integration tests were running "bash tools/rust/*.sh" scripts
  when asked to run Rust tools

- the shell script were running integration tools from ~/.cargo/bin

This is bad because the developers can accidentally install tools
into their home directory.

Now we have the following hack:

- "rustthemis_integration_tools" target defined in test.mk file
  calls Cargo to install integration testing tools into Themis
  build directory ("build" at repository root by default)

- the tools are then copied to tools/rust with "*.rust" extension

- integration tests were run Rust tools as "env tools/rust/*.rust"
  pretending that they are actually scripts

This is still kinda bad because we pollute the tools/rust directory
with throwaway binaries. However, integration tests already do this
to tests/_integration directory with integration driver scripts, so
adding some more binaries should be acceptable.

(Oh, and the keygen tool has been officially renamed into keygen_tool
to avoid conflicts with docs/examples/rust/keygen.rs)

* Clean Rust build files on "make clean"

Cargo keeps its build directory in "target". This can be changed, but
it is very inconvenient for developers.

Nevertheless, running "make clean" should ensure a clean build so run
"cargo clean" after removing regular build files. Also remove the
temporary binaries which may have been copied into tools/rust.
Do this only if Rust toolchain is installed in the system.