Panic on empty input for Secure Cell
It does not make much sense to use an empty master key for encryption,
or encrypt or decrypt empty messages. Secure Cell does not allow such
inputs and returns errors if the user insists. Additionally, context
imprint mode of Secure Cell requires a non-empty user context.

The programmers should never really use Secure Cell in this way. If an
empty key, or message, or token, or user context is used anywhere then
it's most likely to be a programming error. Therefore treat it as such
-- like using an invalid index with a slice -- and panic instead of
returning an error code if we encounter empty inputs.