Do not use fixed user context with Secure Cell (#358)
* Do not use fixed user context with Secure Cell

It is really unwise to have user context the same for all messages
handled by Secure Cell. The idea behind context is for it to be
associated with message content, but distinct and separable from it.
Therefore the context has be different for each message, not the
same for all of them.

With this in mind, drop SecureCell::with_key_and_context() constructor
and introduce encrypt_with_context() and decrypt_with_context() methods
for Secure Cell operation modes. These methods accept the user context
as their first argument. Regular encrypt() and decrypt() methods are
equivalent to using an empty context.

Note that Secure Cell always requires non-empty context in context
imprint mode thus this flavor has only *_with_context() methods.

While we're here, seize this chance to replace ad-hoc generics in
Secure Cell methods with "impl Trait" feature available in Rust 2018.
This improves readability, especially for token-protect mode. However,
this is also a breaking change as impl Trait does not allow explicit
disambiguation.

Finally, with these changes we no longer need KeyBytes::empty()
constructor. Drop it.

* Consistent names of Secure Message methods

It does not really make sense to keep the wrap/unwrap naming scheme
since Secure Message API in Rust is different from the one in C.
Let's rename the methods to be consistent with documentation.

While we're here, update the methods to use impl Trait just like
we did before with Secure Cell.

* Mention breaking changes in changelog

* Accept optional user context as last parameter

Use the "message(, token)?(, user context)?" order for consistency with
other language wrappers. Inputs are not strongly typed so it's very easy
to supply them in a different order, which is very easy to do if you're
used to a particular order from a different language.