Avoid overflow in Secure Cell decryption in sealed mode
Corrupted secure message header or truncated message can trigger an
overflow in message length computation. This integer overflow results
in a segmentation fault later. Handle the overflow and return an error.

We do not have enough information to analyze the header and determine
whether it is corrupted or not. So in case it is corrupted we still
compute a (garbage) message length and then fail decryption if the
provided buffer length is not enough.

Add some tests for Secure Cell to verify authentication context
corruption. Only sealing and token protect mode have authentication
context, there are no tests for context imprint mode.