Check IV and salt length for symmetric encryption
If the length values are corrupted or incorrect it may trigger
a segmentation fault in cryptographic backend. Both OpenSSL and
BoringSSL are affected.

We should check the length of initialization vector before passing
it to EVP_EncryptInit_ex() or EVP_DecryptInit_ex() because those
functions do not perform any length checks. They expect the buffers
to have suitable lengths (or else they may corrupt the stack).

These functions do not check the length of keys as well, but we
do those checks during key derivation.

We should also check that salt length is zero if there is no salt
because PKCS5_PBKDF2_HMAC() does not check that and may crash due
to NULL dereference.

Add some tests to verify this behavior. Encryption and decryption are
implemented via a common function so it is enough to check only one of
them.