Enfore non-empty keys in KeyBytes
It does not really make sense to have empty key material so let's
enforce this restriction globally. Add length checks to all KeyBytes
constructors and have them return Result which can be an error if
the provided byte slice is empty.

from_vec() constructors are used internally to wrap key bytes that
we receive from key generation functions. These should never be
empty so it's okay to panic if they happen to be empty.

try_from_slice() constructors are intended for users. They can pass
empty slices here, but we do check for that and return an error.
The API already uses result and get_key_kind() already performs
length checks so there are no breaking changes here.